% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCallPrice}
\alias{getCallPrice}
\title{Runs MC simulation}
\usage{
getCallPrice(
  Expiry = 0.5,
  Strike = 100,
  Spot = 120,
  Vol = 0.2,
  r = 0.06,
  barrier = 50,
  NumberOfPaths = 10000L
)
}
\arguments{
\item{Expiry}{Time to maturity}

\item{Strike}{strike price}

\item{Spot}{price of the underlying at the moment of option pricing}

\item{Vol}{annualized volatility rate}

\item{r}{annualized risk-free rate}

\item{barrier}{barrier level}

\item{NumberOfPaths}{number of simulations}
}
\value{
option price
}
\description{
Runs MC simulation
}
